+++ packages/web/CHANGELOG.md
@@ -1,5 +1,33 @@
 # posthog-js-lite
 
+## 4.6.48
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @posthog/core@1.30.9
+
+## 4.6.47
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @posthog/core@1.30.8
+
+## 4.6.46
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @posthog/core@1.30.7
+
+## 4.6.45
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @posthog/core@1.30.6
+
 ## 4.6.44
 
 ### Patch Changes